pyproject.toml
[project]
name = "transcript"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.14.0",
    "moviepy==2.2.1",
    "pydub==0.25.1",
    "openai-whisper==20250625",
    "torch==2.9.1",
    "audioop-lts",
]

[project.scripts]
transcript = "transcript.transcribe_video:main"

[tool.setuptools.packages.find]
where = ["src"]